Fixed controllers

+++ Practice_team06/Services/IHallService.cs
@@ -5,7 +5,7 @@ namespace Practice_team06.Services;
 public interface IHallService
 {
     Task<IEnumerable<HallDto>> GetAllAsync();
-    Task<HallDto?> GetByIdAsync(short id);
+    Task<HallDto?> GetByIdAsync(int id);
     Task<HallDto> CreateAsync(CreateHallDto dto);
     Task<bool> DeleteAsync(short id);
     Task<int> GenerateStandardSeatsAsync(GenerateStandardSeatsDto dto);
